Split new component classes into files
